requirements.txt
# for the game
pygame-ce

# for loading the tilemap
pytmx